+++ examples/robust_phase_estimation.py
@@ -15,11 +15,11 @@
 # %% [markdown]
 # # Robust Phase Estimation
 #
-# This example introduces the Robust Phase Estimation algorithm; this QPE version requires only a single ancilla/phase qubit. The way the algorithm is implemented is inspired from J.Gunther et al., *Phase estimation with partially randomized time evolution* [arxiv:2503.05647](https://arxiv.org/abs/2503.05647).
+# This example introduces the Robust Phase Estimation algorithm; this QPE version requires only a single ancilla/phase qubit. Our implementation is inspired by [J. Günther et al., Phase Estimation with Partially Randomized Time Evolution](https://doi.org/10.1103/ynxb-p2xq).
 #
 # In this notebook we explain the idea of the algorithm and apply it to simple models: the Heisenberg model with $4$ spins, the H$_2$ molecule in the minimal basis.
 #
-# We study the Trotter and statistical errors, and check that the RPE algorithm verifies Heisenberg scaling, i.e. the possibility to measure the energy with precision $\varepsilon$ in time $\mathcal{O}(1/\varepsilon)$.
+# We study the Trotter and statistical errors, and check that the RPE algorithm satisfies Heisenberg scaling, i.e. the ability to measure the energy with precision $\varepsilon$ in time $\mathcal{O}(1/\varepsilon)$.
 
 # %%
 import time
@@ -42,10 +42,10 @@ from qpe_toolbox.hamiltonian import (
 plt.rcParams.update({"font.size": 12})
 
 # %% [markdown]
-# ## Hadamard test
+# ## Hadamard Test
 # The Robust Phase Estimation algorithm relies on the Hadamard test procedure, which we introduce below. Our presentation takes inspiration from [Lin Lin's lecture notes](https://math.berkeley.edu/~linlin/qasc/) and the ["Hadamard test" Wikipedia page](https://en.wikipedia.org/wiki/Hadamard_test).
 #
-# The goal of the Hadamard test is to compute $\bra{\psi} U \ket{\psi}$ where $U$ is a unitary operator. Since $U$ is generally not Hermitian, it is not an observable, therefore the real and imaginary part of $\bra{\psi} U \ket{\psi}$ must be measured separately.
+# The goal of the Hadamard test is to compute $\bra{\psi} U \ket{\psi}$ where $U$ is a unitary operator. Since $U$ is generally not Hermitian, it is not an observable; therefore the real and imaginary parts of $\bra{\psi} U \ket{\psi}$ must be measured separately.
 #
 # The idea is to build a random variable whose expectation value gives the real (resp. imaginary) part of $\bra{\psi} U \ket{\psi}$.
 # Consider the following circuit:
@@ -54,10 +54,10 @@ plt.rcParams.update({"font.size": 12})
 #
 # *Taken from Günther et al. arXiv:2503.05647*.
 #
-# The Hadamard test uses a single auxiliary qubit initially in state $\ket{0}$ and a data register with $n$ qubits initialized in state $\ket{\psi}$.
-# We start by applying the the Hadamard gate $H$ to the auxiliary qubit to put it in a superposition state. Then we apply a controlled-$U$ gate to the data register conditioned on the auxiliary qubit, followed by a a rotation (PHASE) gate  $R(\theta)$ and finally another Hadamard gate on the control qubit.
+# The Hadamard test uses a single auxiliary qubit initially in state $\ket{0}$ and a physical register with $n$ qubits initialized in state $\ket{\psi}$.
+# We start by applying the Hadamard gate $H$ to the auxiliary qubit to put it in a superposition state. Then we apply a controlled-$U$ gate to the physical register conditioned on the auxiliary qubit, followed by a rotation (PHASE) gate  $R(\theta)$ and finally another Hadamard gate on the control qubit.
 #
-# At the end of the circuit we measure the control qubit and define a random variable $\textbf{Z}_\theta$: if the result of the measure is $\ket{0}$, we output $1$, if the result is $\ket{1}$ we output $-1$. The expectation value of $\textbf{Z}_\theta$ satisfies:
+# At the end of the circuit we measure the control qubit and define a random variable $\textbf{Z}_\theta$: if the result of the measurement is $\ket{0}$, we output $1$, if the result is $\ket{1}$, we output $-1$. The expectation value of $\textbf{Z}_\theta$ satisfies:
 #
 # $$ \mathbb{E}\textbf{Z}_\theta = P(0) - P(1) = \mathrm{Re} \left(e^{i\theta} \bra{\psi}U\ket{\psi}\right). $$
 #
@@ -114,13 +114,12 @@ E0, psi0 = do_dmrg(H)
 #
 # The function `run_hadamard_test` runs the Hadamard test and returns $\mathrm{Re}~e^{i\theta} \bra{\psi}U\ket{\psi}$.
 #
-# Below we estimate $E_0$ by running the Hadamard test for a time evolution during a random time $t$.
+# Below we estimate $E_0$ by running the Hadamard test for time evolution over a given time $t$.
 #
 # We first consider exact time evolution.
 
 # %%
-rng = np.random.default_rng(seed=42)
-t = rng.random()
+t = 0.7  # exact value does not matter here
 exact_unitary = H.get_U_exact(t)
 
 n_shots = EXACT  # exact computation (no sampling)
@@ -156,14 +155,15 @@ print(f"error = {abs(np.angle(Z) / t + E0):.2g}")
 #
 
 # %%
-shots_list = np.array([10, 100, 500, 1000, 2000])
+shots_list = np.array([10, 50, 100, 200, 400, 500, 800, 1000, 1500, 2000])
 errors = []
 durations = []
+rng = np.random.default_rng(42)
 
 for n_shots in tqdm.tqdm(shots_list):
     st = time.time()
-    X = qpe.run_hadamard_test(psi0, exact_unitary, 0, n_shots=n_shots)
-    Y = qpe.run_hadamard_test(psi0, exact_unitary, -np.pi / 2, n_shots=n_shots)
+    X = qpe.run_hadamard_test(psi0, exact_unitary, 0, n_shots, rng=rng)
+    Y = qpe.run_hadamard_test(psi0, exact_unitary, -np.pi / 2, n_shots, rng=rng)
     et = time.time() - st
 
     Z = X + 1j * Y
@@ -173,46 +173,57 @@ for n_shots in tqdm.tqdm(shots_list):
     durations.append(et)
 
 # %% [markdown]
-# The statistical error decreases as $1/\sqrt{N_{\rm shots}}$ while the computation time increases linearly with $N_{\rm shots}$
+# The statistical error decreases as $1/\sqrt{N_{\rm shots}}$ while the computation time increases linearly with $N_{\rm shots}$.
 
 # %%
+log_prefactor = np.log(errors) + np.log(shots_list) / 2
+mean = np.exp(log_prefactor.mean())
+std = np.exp(log_prefactor.std())
 fig, (ax_e, ax_t) = plt.subplots(nrows=2)
 fig.subplots_adjust(hspace=0.4)
-ax_e.loglog(shots_list, errors, "-o")
 ax_e.loglog(
     shots_list,
-    errors[0] * np.sqrt(10 / shots_list),
+    mean / np.sqrt(shots_list),
     "--",
     label="$\\propto 1/\\sqrt{N_{\\rm shots}}$",
-    zorder=0,
 )
+ax_e.fill_between(
+    shots_list,
+    mean / std / np.sqrt(shots_list),
+    mean * std / np.sqrt(shots_list),
+    alpha=0.2,
+)
+ax_e.loglog(shots_list, errors, "-o")
+
 ax_t.plot(shots_list, durations, "-o")
-ax_e.legend()
+ax_e.legend(loc="lower left")
 ax_e.set_xlabel("number of shots")
 ax_t.set_xlabel("number of shots")
 ax_e.set_ylabel("error (units of J)")
 ax_t.set_ylabel("duration (seconds)");
 
 # %% [markdown]
-# ## Robust phase estimation algorithm
+# ## Robust Phase Estimation Algorithm
 #
 # ### Introduction
 #
-# Quote from Günther et al. [arxiv:2503.05647](https://arxiv.org/abs/2503.05647):
+# In the previous section, we defined a function $g(t) = \mathbb{E}[\mathbf{Z}(t)] = \bra{\psi} U(t) \ket{\psi}$ is computed using the Hadamard test. In this section, we explain how to process this function within the Quantum Phase Estimation algorithm.
+#
+# Quote from Günther et al. [PRX Quantum 7, 020332](https://doi.org/10.1103/ynxb-p2xq):
 # > "If we think of g(t) as a time signal, then the phase estimation routine will constitute a signal processing transformation to compute the lowest frequency of $g(t)$ (corresponding to the energy $E_0$), provided that we have some guarantee on the overlap of $\ket{\psi}$ with the ground state; we assume a lower bound
 #   $c_0 \geq \eta$. With appropriate signal processing methods, one can find the value of $E_0$ with accuracy $\varepsilon$ using $M$ circuits with time evolution for
 #  times $t_1, . . . , t_M$. This can be done such that the maximal time evolution $t_{\rm max} = \mathrm{max}\{t_1, . . . , t_M\}$ and the total time over all circuit runs $t_{\rm tot} = t_1 + t_2 + · · · + t_M$  both scale as $\varepsilon^{-1}$. This Heisenberg scaling is known to be optimal."
 #
-# ### Algorithm
+# ### Description of the Algorithm
 # To get precision $\varepsilon$, the idea of the robust phase estimation algorithm is to consider $M$ different circuits where $M = \lceil \log_2 \varepsilon^{-1} \rceil$ and estimate
 #
-# $$ g(2^m) = \mathbb{E}[Z(2^{m})] = \exp(-i 2^{m} E_0) $$
+# $$ g(2^m) = \mathbb{E}[\mathbf{Z}(2^{m})] = \exp(-i 2^{m} E_0) $$
 #
-# for $m=0,1,..,M-1$. Each iteration gives a supplementary bit of precision on $E_0$.
+# for $m=0,1,..,M-1$. At each step the algorithm computes an estimate $\theta_m$ of $E_0$; the objective is that $\theta_m$ be the best $m$-bit approximation of $E_0$, so that each iteration adds one bit of precision.
 #
-# **Algorithm 1 (p.24)**
+# **Algorithm 1 (arXiv p.24 / PRX p.20)**
 #
-# 1. The algorithm is initialized with $\theta_{-1}=0$ so that $\theta_0 = \phi_0$.
+# 1. The algorithm is initialized with $\theta_{-1}=0$.
 #
 # 2. For each $m$:
 #
@@ -221,99 +232,88 @@ ax_t.set_ylabel("duration (seconds)");
 #     $$ \bar{\bf Z}(2^{m}) = \frac{1}{N_{\rm shots}} \sum_{n=1}^{N_{\rm shots}} {\bf Z}^{(n)} (2^m) $$
 #
 #
-#     - From the outcome we compute $\phi_m = - \arg(\bar{Z}(2^m))$ using `numpy.angle`
+#     - From the outcome we compute $\phi_m = - \arg(\bar{Z}(2^m))$
 #
-#     - By definition $\phi_m \in ]-\pi, \pi]:~\phi_m~$  is an approximation of $2^mE_0$ modulo $2\pi$
+#     - By definition $\phi_m \in (-\pi, \pi]:~\phi_m~$  is an approximation of $2^mE_0$ modulo $2\pi$
 #
 #     - Given a previous guess $\theta_{m-1}$ for the ground state energy $E_0$, the new energy estimate $\theta_m$ is given by
 #
 #     $$ \theta_m = 2^{-m} (2\pi k + \phi_m), $$
-#     where $k$ is an integer between $0$ and $2^m - 1$ which minimizes the distance
+#     where $k$ is an integer between $0$ and $2^m - 1$ chosen such that $\theta_m $ minimizes the distance
 #
-#      $$ d(\theta_m, \theta_{m-1}) = \min_{q\in\mathbb{Z}} | \theta_m - \theta_{m-1} + 2q\pi|, $$
-#       under the condition $-\pi < \theta \leq \pi$.
+#      $$ d(\theta_m, \theta_{m-1}) = \min_{q\in\mathbb{Z}} | \theta_m - \theta_{m-1} + 2\pi q|, $$
+#       under the condition $-\pi < \theta_m \leq \pi$.
 #
-# The algorithm ensures that at each step, $\theta_m$ is the best $m$-bit approximation of $E_0$.
+# The algorithm ensures that at each step, $\theta_m$ is indeed the best $m$-bit approximation of $E_0$.
 # The following lemma guarantees convergence:
 #
-# **Lemma B.1. (p.25)**:
-# if $d(\phi_k,2^{k}E)<\frac{\pi}3$ for $k=0,1,...,m$ then $\theta_m$ is such that $d(\theta_m,E) \leq 2^{-m}\frac{\pi}3$
+# **Lemma B.1. (arXiv p.25 / PRX p.21)**:
+# if $d(\phi_m,2^{m}E)<\frac{\pi}3$ for $m=0,1,...,M -1$ then $\theta_m$ is such that $d(\theta_m,E) \leq 2^{-m}\frac{\pi}3$
 
 # %% [markdown]
-# #### Illustration of the distance $d(\theta,\phi)$
-#
-# To build an intuition, let us plot the distance $d(\theta, \phi)$ as a function of $\theta$ for a given $\phi$ ($ 3\pi/2$ in the example) and vice-versa (the distance is symmetric by definition).
+# To build an intuition, let us plot the distance $d(\theta, \phi)$ as a function of $\theta$ for a given $\phi$ ($ \phi = 3\pi/2$ in the example) and vice-versa (the distance is symmetric by definition).
 
 # %%
 thetas = np.linspace(-2 * np.pi, 2 * np.pi, 600)
 plt.xticks(
-    [i * np.pi for i in range(-2, 3)],
-    [r"$-2\pi$", r"$-\pi$", "$0$", r"$\pi$", r"$2\pi$"],
+    np.pi * np.arange(-2, 3), [r"$-2\pi$", r"$-\pi$", "$0$", r"$\pi$", r"$2\pi$"]
 )
-plt.yticks([0, np.pi / 2, np.pi], ["0", "$\\pi/2$", "$\\pi$"])
-
+plt.yticks([0, np.pi / 2, np.pi], ["0", r"$\pi/2$", r"$\pi$"])
 plt.xlabel(r"$\theta$")
 
 plt.plot(
-    thetas,
-    [qpe.rpe_distance(t, 3 * np.pi / 2) for t in thetas],
-    label=r"$d(3\pi/2, \theta)$",
+    thetas, qpe.angular_distance(thetas, 3 * np.pi / 2), label=r"$d(3\pi/2, \theta)$"
 )
 plt.plot(
     thetas,
-    [qpe.rpe_distance(3 * np.pi / 2, t) for t in thetas],
+    qpe.angular_distance(3 * np.pi / 2, thetas),
     "--",
     label=r"$d(\theta, 3\pi/2)$",
 )
 plt.legend();
 
 # %% [markdown]
-# Let us now illustrate the first two steps of the algorithm for concreteness. We start with $N_{\rm shots}=2$ and $\epsilon = 0.2$. We also consider exact time evolution.
+# Let us now illustrate the first two steps of the algorithm for concreteness. We start with $N_{\rm shots}=2$ and consider exact time evolution.
 
 # %%
-sign_E0 = np.sign(E0)
-
-epsilon = 0.02
-M = int(np.ceil(np.log2(1 / epsilon)))
-
 n_shots = 2
+rng = np.random.default_rng(42)
 
 # m = 0
-phi_0 = qpe.rpe_get_hadamard_output(H, psi0, 0, EXACT, n_shots)
+phi_0 = qpe.rpe_get_hadamard_output(H, psi0, 0, EXACT, n_shots, rng=rng)
 theta_0 = phi_0
 
 m = 1
-phi_1 = qpe.rpe_get_hadamard_output(H, psi0, m, EXACT, n_shots)
-S_1 = [(phi_1 + sign_E0 * 2 * np.pi * k) / 2**m for k in range(2**m)]
+phi_1 = qpe.rpe_get_hadamard_output(H, psi0, m, EXACT, n_shots, rng=rng)
+# candidate energies, wrapped into (-pi, pi]
+possible_phases = (phi_1 + 2 * np.pi * np.arange(2**m)) / 2**m
+possible_phases = (possible_phases + np.pi) % (2 * np.pi) - np.pi
 
 # %% [markdown]
-# Let's visualize how the different elements of $S_1$ compare to $\theta_0$
+# Let us visualize how the different possible phases compare to $\theta_0$
 
 # %%
-plt.hlines(1, -np.pi, np.pi, "k")
-plt.plot([-np.pi, np.pi], [1, 1], "|", markersize=10, color="k")
+plt.plot([-np.pi, np.pi], [1, 1], marker="|", markersize=10, color="k")
 plt.plot(theta_0, 1, "*", markersize=20, color="r", label=r"$\theta_0$")
-y = np.ones(np.shape(S_1))
-plt.plot(S_1, y, "o", markersize=15, label=r"$S_1$")
+plt.plot(possible_phases, [1, 1], "o", markersize=15, label=r"possible $\theta_1$")
 
 plt.text(-1.1 * np.pi, 0.99, r"$-\pi$", fontsize=16)
 plt.text(np.pi, 0.99, r"$\pi$", fontsize=16)
 plt.text(1.05 * theta_0, 0.99, r"$\theta_0$", fontsize=16, color="r")
-
 plt.axis("off")
 plt.legend();
 
 # %% [markdown]
-# We compute $\theta_1$ as the element from $S_1$ closest to $\theta_0$ and check that the error decreases between the first and second iteration:
+# We compute $\theta_1$ as the closest possible phase to $\theta_0$ and check that the error decreases between the first and second iteration:
 
 # %%
-theta_1, d_min = qpe.rpe_update_theta(S_1, theta_0)
-print(f"Exact energy E = {E0:.4f}")
-print(f"theta_0 = {theta_0:.4f}")
-print(f"theta_1 = {theta_1:.4f}")
+theta_1 = qpe.rpe_update_theta(phi_1, theta_0, m)
+print(f"Exact energy {E0 = :.4f}")
+print(f"{theta_0 = :.4f},   error = {abs(E0 - theta_0):.4f}")
+print(f"{theta_1 = :.4f},   error = {abs(E0 - theta_1):.4f}")
 
 # %% [markdown]
-# ### Run RPE. Statistical precision
+# ### Statistical Precision
 #
 # We will now run the algorithm and see the influence of statistical noise. The `robust_phase_estimation` function returns the full list of $\theta_m$, $m=0,...,M-1$.
 #
@@ -326,49 +326,49 @@ print(f"Target precision epsilon={epsilon}: requires M={M} iterations\n")
 
 n_shots = 1
 
-theta_list = qpe.robust_phase_estimation(
-    H, psi0, epsilon, sign_E0, EXACT, n_shots, verbosity=1
+rng = np.random.default_rng(42)
+theta_values = qpe.robust_phase_estimation(
+    H, psi0, M, EXACT, n_shots, verbosity=1, rng=rng
 )
 
 # %%
 plt.semilogy(
-    [qpe.rpe_distance(theta, E0) for theta in theta_list[1:]],
-    "-o",
-    label=f"$N_{{\\rm shots}}={n_shots}$",
+    qpe.angular_distance(theta_values, E0), "-o", label=f"$N_{{\\rm shots}}={n_shots}$"
 )
-plt.semilogy([np.pi / 3 / 2**i for i in range(M + 1)], "k--", label="$2^{-m}~\\pi/3$")
+plt.semilogy(np.pi / 3 / 2 ** np.arange(M), "k--", label="$2^{-m}~\\pi/3$")
 plt.legend()
 plt.xlabel("iteration $m$")
 plt.ylabel("$d(\\theta_m, E)$");
 
 # %% [markdown]
-# We see that for such small systems, with exact time evolution, a single shot gives an estimate that converges. We now increase the number of shots to improve the precision
+# We see that for such small systems, with exact time evolution, a single shot gives an estimate that converges. We now increase the number of shots to improve the precision.
 
 # %%
 n_shot_list = [1, 2, 3, 4]
 
+rng = np.random.default_rng(42)
 for n_shots in n_shot_list:
-    theta_list = qpe.robust_phase_estimation(H, psi0, epsilon, sign_E0, EXACT, n_shots)
+    theta_values = qpe.robust_phase_estimation(H, psi0, M, EXACT, n_shots, rng=rng)
     plt.semilogy(
-        [qpe.rpe_distance(theta, E0) for theta in theta_list[1:]],
+        qpe.angular_distance(theta_values, E0),
         "-o",
         label=f"$N_{{\\rm shots}}={n_shots}$",
     )
 
-plt.semilogy([np.pi / 3 / 2**i for i in range(M + 1)], "k--", label="$2^{-m}~\\pi/3$")
+plt.semilogy(np.pi / 3 / 2 ** np.arange(M), "k--", label="$2^{-m}~\\pi/3$")
 plt.legend()
 plt.xlabel("iteration $m$")
 plt.ylabel("$d(\\theta_m, E)$")
 plt.title(rf"$\epsilon={epsilon},\; M={M}$");
 
 # %% [markdown]
-# ### Trotter approximation of the time evolution operator
+# ### RPE with Trotter Approximation
 #
 # We now apply the same algorithm but replace the exact time evolution operator by a second order Trotter approximation.
 #
 # The `n_trotter_steps` argument in the `robust_phase_estimation` function sets the number of Trotter steps for $m=0$. The number of steps is multiplied by $2$ at each iteration to keep the Trotter timestep constant.
 #
-# The computation will now take longer since the number of gates for the time evolution now grows like $2^m$.
+# The computation will now take longer since the number of gates for the time evolution grows like $2^m$.
 
 # %%
 # %%time
@@ -378,19 +378,20 @@ n_trotter_steps = 1
 n_shots = 4
 thetas_ttr_list = []
 
+rng = np.random.default_rng(42)
 thetas_ttr = qpe.robust_phase_estimation(
-    H, psi0, epsilon, sign_E0, n_trotter_steps, n_shots, verbosity=1
+    H, psi0, M, n_trotter_steps, n_shots, trotter_order=2, verbosity=1, rng=rng
 )
 
 thetas_ttr_list.append(thetas_ttr)
 
 # %%
 plt.semilogy(
-    [qpe.rpe_distance(theta, E0) for theta in thetas_ttr_list[0][1:]],
+    qpe.angular_distance(thetas_ttr_list[0], E0),
     "-o",
     label=f"$n_{{\\rm steps}}={n_trotter_steps}$",
 )
-plt.semilogy([np.pi / 3 / 2**i for i in range(M + 1)], "k--", label="$2^{-m}~\\pi/3$")
+plt.semilogy(np.pi / 3 / 2 ** np.arange(M), "k--", label="$2^{-m}~\\pi/3$")
 plt.legend()
 plt.title(r"$N_{\rm shots}=4$")
 plt.xlabel("iteration $m$")
@@ -403,76 +404,71 @@ plt.ylabel("error");
 # %%time
 
 n_trotter_steps = 2
+rng = np.random.default_rng(42)
 thetas_ttr = qpe.robust_phase_estimation(
-    H, psi0, epsilon, sign_E0, n_trotter_steps, n_shots, verbosity=1
+    H, psi0, M, n_trotter_steps, n_shots, trotter_order=2, verbosity=1, rng=rng
 )
 thetas_ttr_list.append(thetas_ttr)
 
 # %%
 for i, n_trotter_steps in enumerate([1, 2]):
     plt.semilogy(
-        [qpe.rpe_distance(theta, E0) for theta in thetas_ttr_list[i][1:]],
+        qpe.angular_distance(thetas_ttr_list[i], E0),
         "-o",
         label=f"$n_{{\\rm steps}}={n_trotter_steps}$",
     )
 
-plt.semilogy([np.pi / 3 / 2**i for i in range(M + 1)], "k--", label="$2^{-m}~\\pi/3$")
+plt.semilogy(np.pi / 3 / 2 ** np.arange(M), "k--", label="$2^{-m}~\\pi/3$")
 plt.legend()
 plt.text(4.5, 0.3, "$N_{\\rm shots}=4$")
 plt.xlabel("iteration $m$")
 plt.ylabel("error");
 
 # %% [markdown]
-# ## Heisenberg scaling
+# ## Heisenberg Scaling
 #
-# The experimental time is proportional to $N_{\rm shots} \sum_{m=0}^M  2^m$, i.e. it scales like $2^M$. The RPE algorithm reaches a precision $\varepsilon$ in $M = \lceil \log_2 \varepsilon^{-1} \rceil$ iterations.
+# The experimental time is proportional to $N_{\rm shots} \cdot \sum_{m=0}^{M-1}  2^m$, i.e. it scales like $2^M$. With $M = \lceil \log_2 \varepsilon^{-1} \rceil$ iterations, the RPE algorithm reaches a precision of order $\varepsilon$ (the guaranteed bound is $2^{-(M-1)}\pi/3 \simeq 2\varepsilon$).
 # Hence it achieves Heisenberg scaling: reaching a precision $\varepsilon$ in time $\mathcal{O}(2^M) = \mathcal{O}(1/\varepsilon)$.
 #
-# Let us illustrate that below: we run the RPE algorithm for various $\varepsilon$ and plot experimental time versus energy error. (We take exact time evolution for simplicity: in this case we consider that the experimental time is exactly $N_{\rm shots} \sum_{m=0}^M  2^m$.)
+# Let us illustrate that below: we run the RPE algorithm for various $\varepsilon$ and plot experimental time versus energy error. Here we use exact time evolution for simplicity: in this case the experimental time is exactly $N_{\rm shots} \cdot \sum_{m=0}^{M-1}  2^m$.
 
 # %%
-epsilon_list = [0.1 / 2**m for m in range(11)]
+epsilon_values = 0.1 / 2 ** np.arange(11)
 n_shots = 5
 
 cost_list = []
 res_list = []
-for epsilon in epsilon_list:
+rng = np.random.default_rng(42)
+for epsilon in epsilon_values:
     M = int(np.ceil(np.log2(1 / epsilon)))
-    cost_list.append(sum([n_shots * 2**m for m in range(M + 1)]))
-    theta_list = qpe.robust_phase_estimation(H, psi0, epsilon, sign_E0, EXACT, n_shots)
+    cost_list.append(sum([n_shots * 2**m for m in range(M)]))
+    theta_list = qpe.robust_phase_estimation(H, psi0, M, EXACT, n_shots, rng=rng)
     res_list.append(theta_list[-1])
 
 # %%
-plt.loglog(cost_list, [abs(en - E0) for en in res_list], "-o", label="RPE")
-plt.loglog(cost_list, epsilon_list, "rd", label=r"target $\epsilon$")
+xfit = np.array([0.2, 0.1 / 2**15])
+plt.loglog(xfit**-2, xfit, "k:", label=r"$t_{tot}=1/\epsilon^2$")
+plt.loglog(xfit**-1, xfit, "b:", label=r"$t_{tot}=1/\epsilon$")
+plt.loglog(cost_list, epsilon_values, "rd", label=r"target $\epsilon$")
 plt.loglog(
-    [1 / eps**2 for eps in epsilon_list],
-    epsilon_list,
-    "k:",
-    label=r"$t_{tot}=1/\epsilon^2$",
-)
-plt.loglog(
-    [1 / eps for eps in epsilon_list], epsilon_list, "b:", label=r"$t_{tot}=1/\epsilon$"
+    cost_list, [qpe.angular_distance(en, E0) for en in res_list], "-o", label="RPE"
 )
 
 plt.xlabel("Experimental time $t_{tot}$ (number of shots x repetitions)")
 plt.ylabel(r"Energy error $\epsilon$")
 plt.title("Heisenberg scaling of RPE with exact time evolution")
-plt.text(10, 5e-5, "Heisenberg Hamiltonian \n4 spins, $S=1/2$, $J=1$")
-plt.legend();
+plt.xlim(10, 2e5)
+plt.legend(title="Heisenberg Hamiltonian\n4 spins, $S=1/2$, $J=1$");
 
 # %% [markdown]
-# ## Quantum chemistry example: diatomic Hydrogen
+# ## Quantum Chemistry Example: Diatomic Hydrogen
 
 # %% [markdown]
 # Let us now consider a molecule: we take $H_2$ in the minimal atomic orbital basis STO-3G. The Hamiltonian in qubit form is obtained via a Jordan-Wigner transformation.
-# Contrary to the previous spin Hamiltonian, the molecular Hamiltonian is non-local: it couples qubits at long distance.
+# Unlike the previous spin Hamiltonian, the molecular Hamiltonian is non-local: it couples distant qubits.
 
 # %%
-mol = gto.M(
-    atom=[("H", (0.0, 0.0, 0.0)), ("H", (0.0, 0.0, 0.735))],
-    basis="STO-3G",
-)
+mol = gto.M(atom=[("H", (0.0, 0.0, 0.0)), ("H", (0.0, 0.0, 0.735))], basis="STO-3G")
 
 H_H2 = chemistry_hamiltonian(
     mol, hf_mode="rhf", encoding="original", do_fci=True, do_ccsd=False
@@ -481,34 +477,33 @@ E0_H2, psi0_H2 = do_dmrg(H_H2)
 print(f"E_DMRG : {E0_H2 + H_H2.e_const:.10f}")
 
 # %% [markdown]
-# ### Exact time evolution
+# ### Exact Time Evolution
 
 # %% [markdown]
-# The system is small enough for exact exponentiation of the Hamiltonian matrix, and exact time evolution
+# The system is small enough for exact exponentiation of the Hamiltonian matrix, and exact time evolution.
 
 # %%
 epsilon = 0.02
 M = int(np.ceil(np.log2(1 / epsilon)))
-sign_E0 = np.sign(E0_H2)
-n_shot_list = [2, 3, 4]
+n_shot_list = [2, 3, 1000]
 
+# n_shot stays small, result depends a lot on seed
+rng = np.random.default_rng(2)
 for n_shots in n_shot_list:
-    theta_list = qpe.robust_phase_estimation(
-        H_H2, psi0_H2, epsilon, sign_E0, EXACT, n_shots
+    theta_values = qpe.robust_phase_estimation(
+        H_H2, psi0_H2, M, EXACT, n_shots, rng=rng
     )
-    plt.semilogy(
-        [qpe.rpe_distance(theta, E0_H2) for theta in theta_list[1:]],
-        "-o",
-        label=f"$N_{{\\rm shots}}={n_shots}$",
-    )
+    distances = qpe.angular_distance(theta_values, E0_H2)
+    plt.semilogy(distances, "-o", label=f"$N_{{\\rm shots}}={n_shots}$")
 
-plt.semilogy([np.pi / 3 / 2**i for i in range(M + 1)], "k--", label="$2^{-m}~\\pi/3$")
+plt.semilogy(np.pi / 3 / 2 ** np.arange(M), "k--", label="$2^{-m}~\\pi/3$")
 plt.legend()
+plt.title(r"$H_2$ STO-3G - exact time evolution")
 plt.xlabel("iteration $m$")
 plt.ylabel("$d(\\theta_m, E)$");
 
 # %% [markdown]
-# ### Trotter
+# ### Trotterized Time Evolution
 
 # %% [markdown]
 # The Trotter decomposition of the molecular Hamiltonian is longer, and simulations become harder. The following run will take a minute...
@@ -518,21 +513,18 @@ plt.ylabel("$d(\\theta_m, E)$");
 n_shots = 2
 n_trotter_steps = 1
 
-thetas_ttr = qpe.robust_phase_estimation(
-    H_H2, psi0_H2, epsilon, sign_E0, n_trotter_steps, n_shots, verbosity=1
+rng = np.random.default_rng(42)
+thetas_trotter_H2 = qpe.robust_phase_estimation(
+    H_H2, psi0_H2, M, n_trotter_steps, n_shots, trotter_order=2, verbosity=1, rng=rng
 )
+distances_trotter_H2 = qpe.angular_distance(thetas_trotter_H2, E0_H2)
 
 # %% [markdown]
 # ... and fails to get the desired precision
 
 # %%
-plt.semilogy(
-    [qpe.rpe_distance(theta, E0_H2) for theta in thetas_ttr_list[i][1:]],
-    "-o",
-    label=f"$n_{{\\rm steps}}={n_trotter_steps}$",
-)
-
-plt.semilogy([np.pi / 3 / 2**i for i in range(M + 1)], "k--", label="$2^{-m}~\\pi/3$")
+plt.semilogy(np.pi / 3 / 2 ** np.arange(M), "k--", label="$2^{-m}~\\pi/3$")
+plt.semilogy(distances_trotter_H2, "-o", label=f"$n_{{\\rm steps}}={n_trotter_steps}$")
 plt.legend()
 plt.title(f"$N_{{\\rm shots}}={n_shots}$")
 plt.xlabel("iteration $m$")
@@ -542,9 +534,9 @@ plt.ylabel("error");
 # If you want to reach the expected precision, what would you try to increase first? $N_{\rm shots}$ or $n_{\rm steps}$?
 
 # %% [markdown]
-# ### Chemical accuracy?
+# ### Chemical Accuracy?
 #
-# The standard for chemical accuracy is $\varepsilon = 10^{-3}$ Ha. Hartrees are the default unit in `pyscf`. We can compute directly the number of iterations required for chemical accuracy:
+# The standard for chemical accuracy is $\varepsilon = 10^{-3}$ Ha. Hartrees are the default unit in `pyscf`. We can directly compute the number of iterations required for chemical accuracy:
 
 # %%
 epsilon = 0.001
@@ -553,9 +545,5 @@ print(f"Chemical accuracy eps={epsilon} requires M={M} iterations")
 
 # %% [markdown]
 # Reaching chemical accuracy requires at least ten iterations, and a sufficient number of shots and Trotter steps.
-# If you want to go further, you can first make an estimation of the runtime for $M=10$ and a given number of shots and Trotter steps,
+# If you want to go further, you can first estimate the runtime for $M=10$ and a given number of shots and Trotter steps,
 # then with some patience try to run the simulation.
-
-# %%
-
-# %%